refactor(lang): rename double accessors to float
【动机】
类型名 double 已改 float，内部 Value 层访问器（isDouble/asDouble/toDouble/fromDouble/asDoubleRaw）与 DLL API（make_double/is_double/as_double/register_double）命名需同步，消除 Double 残留。

【原理】
- Value 层驼峰方法：isDouble→isFloat、asDouble→asFloat（含 asDoubleRaw→asFloatRaw）、toDouble→toFloat、fromDouble→fromFloat。
- DLL API 下划线方法：make_double→make_float、is_double→is_float、as_double→as_float、register_double→register_float（含 host_/cb_ 桥接前缀）。
- 保留 JIT 层 IEEE 754 硬件 double（JITType::Double、boxDouble、unboxDouble、createDoubleConstant、GuardIsDouble、Int32ToDouble），因指 64 位 XMM 浮点而非 JC2 的 float 显示名。
- 保留 C++ double 关键字（参数/返回类型）、f64/f32 二进制格式名、listToDoubles 辅助函数。

+++ src/cas/Factorization.cpp
@@ -1018,7 +1018,7 @@ namespace jc {
                         int n = degree;
 
                         auto getExactCos = [](int m, int n_val) -> SymExpr {
-                            int g = static_cast<int>(BigInt::gcd(BigInt(std::abs(m)), BigInt(n_val)).toDouble());
+                            int g = static_cast<int>(BigInt::gcd(BigInt(std::abs(m)), BigInt(n_val)).toFloat());
                             m /= g;
                             n_val /= g;
                             if (m < 0) m = -m;
